+++ BackendChallenge/Dockerfile
@@ -5,25 +5,28 @@ FROM mcr.microsoft.com/dotnet/aspnet:8.0 AS base
 USER $APP_UID
 WORKDIR /app
 EXPOSE 8080
+EXPOSE 8081
 
 
 # This stage is used to build the service project
 FROM mcr.microsoft.com/dotnet/sdk:8.0 AS build
 ARG BUILD_CONFIGURATION=Release
 WORKDIR /src
-COPY ["Backend Challenge/Backend Challenge.csproj", "Backend Challenge/"]
-RUN dotnet restore "./Backend Challenge/Backend Challenge.csproj"
+
+COPY ["BackendChallenge.csproj", "BackendChallenge/"]
+RUN dotnet restore "./BackendChallenge/BackendChallenge.csproj"
+
 COPY . .
-WORKDIR "/src/Backend Challenge"
-RUN dotnet build "./Backend Challenge.csproj" -c $BUILD_CONFIGURATION -o /app/build
+WORKDIR "/src/BackendChallenge"
+RUN dotnet build "./BackendChallenge.csproj" -c $BUILD_CONFIGURATION -o /app/build
 
 # This stage is used to publish the service project to be copied to the final stage
 FROM build AS publish
 ARG BUILD_CONFIGURATION=Release
-RUN dotnet publish "./Backend Challenge.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
+RUN dotnet publish "./BackendChallenge.csproj" -c $BUILD_CONFIGURATION -o /app/publish /p:UseAppHost=false
 
 # This stage is used in production or when running from VS in regular mode (Default when not using the Debug configuration)
 FROM base AS final
 WORKDIR /app
 COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "Backend Challenge.dll"]
+ENTRYPOINT ["dotnet", "BackendChallenge.dll"]
