+++ BackendChallenge/Models/StoryItem.cs
