+++ BackendChallenge/Services/IStoryService.cs
